feat : 새로 만든 RPS.swift에 클래스를 만들어 main.swift에 있던 RockPaperScissors 관련 코드를 이동함.
