+++ RockPaperScissors/RockPaperScissors/main.swift
@@ -17,35 +17,20 @@ while true
 
     do
     {
-        let gameResult : Bool = try playRockPaperScissors()
+        let gameResult : Bool = try RPS().play()
 
         if shouldByClose(gameResult)
         {
             break
         }
+        
+        
     }
     catch RockPaperScissorsError.invalidInput
     {
         print("잘못된 입력입니다. 다시 입력해주세요")
     }
 
-}
-
-func playRockPaperScissors() throws -> Bool
-{
-    let user: Int = try checkInput()
-
-    switch user
-    {
-        case 0:
-            print("게임 종료")
-            return true
-        case 1...3:
-            let gameResult : Bool = getGameResult(of: user)
-            return gameResult
-        default:
-            throw RockPaperScissorsError.invalidInput
-    }
 }
 
 func checkInput() throws -> Int
@@ -55,41 +40,6 @@ func checkInput() throws -> Int
         throw RockPaperScissorsError.invalidInput
     }
     return user
-}
-
-func getGameResult(of user: Int) -> Bool
-{
-    let computer = Int.random(in: 1...3)
-
-    if user == computer
-    {
-        print("비겼습니다")
-        return false
-    }
-    else if isWin(of: user, versus: computer)
-    {
-        print("이겼습니다")
-        return true
-    }
-    else
-    {
-        print("졌습니다")
-        return true
-    }
-}
-
-func isWin(of user: Int, versus computer: Int) -> Bool
-{
-    let winCase: [Int: Int] = [1: 3, 2: 1, 3: 2]
-
-    if winCase[user] == computer
-    {
-        return true
-    }
-    else
-    {
-        return false
-    }
 }
 
 func shouldByClose(_ gameResult : Bool) -> Bool
